requirements.txt
python-ldap
django-auth-ldap
mysql-replication
allure-pytest==2.9.45
allure-python-commons==2.9.45
anyio==3.5.0
appdirs==1.4.4
APScheduler==3.8.1
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
backports.zoneinfo==0.2.1
black==19.10b0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
channels==3.0.3
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
Django==3.1.3
django-apscheduler==0.6.0
django-cors-headers==3.6.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg2==1.19.4
easyocr==1.4.1
Faker==4.18.0
fastapi==0.72.0
h11==0.13.0
httprunner==3.1.6
hyperlink==21.0.0
idna==3.3
imageio==2.9.0
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.9.5
loguru==0.4.1
MarkupSafe==2.0.1
more-itertools==8.12.0
networkx==2.7.1
numpy==1.22.3
openapi-codec==1.3.2
opencv-python==4.1.2.30
opencv-python-headless==4.5.5.64
packaging==21.3
pathspec==0.9.0
Pillow==8.2.0
pluggy==0.13.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
PyJWT==1.7.1
PyMySQL~=0.8.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytesseract==0.3.9
pytest==5.4.3
pytest-assume==2.4.3
pytest-html==2.1.1
pytest-metadata==1.11.0
python-bidi==0.4.2
python-memcached==1.59
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==5.4.1
regex==2022.1.18
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.2
scipy==1.8.0
sentry-sdk==0.14.4
service-identity==21.1.0
simplejson==3.17.6
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.17.1
tep==0.9.6
text-unidecode==1.3
tifffile==2022.2.9
toml==0.10.2
torch==1.10.2
torchvision==0.11.3
txaio==21.2.1
typed-ast==1.5.2
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.4
wcwidth==0.2.5
win32-setctime==1.1.0
wincertstore==0.2
zope.interface==5.4.0
bs4
openpyxl
et-xmlfile
filetype
django-extensions
django-werkzeug-debugger-runserver
tzlocal
six
requests
pycryptodome
matplotlib~=3.5.1
python-dateutil~=2.8.2
dictdiffer~=0.9.0
python-gitlab
channels_redis
